Move query response DTOs into queries for higher cohesion and to avoid reusing DTOs
