--- application/account-management/Core/Users/Queries/GetUsersResponseDto.cs
@@ -1,3 +1,0 @@
-namespace PlatformPlatform.AccountManagement.Core.Users.Queries;
-
-public sealed record GetUsersResponseDto(int TotalCount, int PageSize, int TotalPages, int CurrentPageOffset, UserResponseDto[] Users);
